Unexport `keyCheck` and `keyInfo` from lib/pmcrypto
These functions will be replaced in the future (with breaking API changes), but for now we are interested in keeping the test cases around
