+++ lib/key/utils.js
@@ -9,6 +9,8 @@ import { serverTime } from '../serverTime';
 import { DEFAULT_OFFSET } from '../constants';
 import { getSymmetricKeySize, getRandomBytes } from '../crypto/utils';
 import encryptMessage from '../message/encrypt';
+import { SHA256 } from '../crypto/hash';
+import { arrayToHexString } from '../utils';
 
 /**
  * Generate key and its revocation certificate
@@ -105,6 +107,14 @@ export const canKeyEncrypt = async (publicKey, date = serverTime()) => {
 export function getFingerprint(key) {
     return key.getFingerprint();
 }
+
+export const getSHA256Fingerprints = (key) => {
+    return Promise.all(
+        key.getKeys().map(async ({ keyPacket }) => {
+            return arrayToHexString(await SHA256(keyPacket.writeForHash(keyPacket.version)));
+        })
+    );
+};
 
 /**
  * Find the key entity that generated the given signature.
