+++ lib/pmcrypto.js
@@ -44,8 +44,8 @@ export {
 
 export { default as processMIME } from './message/processMIME';
 
-export { default as keyInfo, getSHA256Fingerprints } from './key/info';
+export { getSHA256Fingerprints } from './key/info';
 
-export { checkKeyStrength, keyCheck } from './key/check';
+export { checkKeyStrength } from './key/check';
 
 export * from './constants';
